Make PerasVoteDB exceptions part of the exposed API
